requirements.txt
asgiref==3.7.2
build==1.0.3
CacheControl==0.13.1
certifi==2024.2.2
charset-normalizer==3.3.2
cleo==2.1.0
colorama==0.4.6
contourpy==1.2.1
crashtest==0.4.1
cycler==0.12.1
distlib==0.3.8
Django==5.0.2
django-crispy-forms==2.1
djangorestframework==3.15.1
dulwich==0.21.7
fastjsonschema==2.19.1
filelock==3.13.1
fonttools==4.51.0
idna==3.6
installer==0.7.0
jaraco.classes==3.3.1
keyring==24.3.0
kiwisolver==1.4.5
matplotlib==3.8.4
more-itertools==10.2.0
msgpack==1.0.7
networkx==3.3
numpy==1.26.4
packaging==23.2
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
ptyprocess==0.7.0
pyparsing==3.1.2
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.2
rapidfuzz==3.6.1
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.4
six==1.16.0
sqlparse==0.4.4
tomlkit==0.12.3
trove-classifiers==2024.2.23
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1
